refactor(cli): extract policy and channels actions
