+++ src/lib/channels-mutate-cli-commands.ts
@@ -5,13 +5,6 @@
 
 import { Args, Command, Flags } from "@oclif/core";
 
-import {
-  addSandboxChannel,
-  removeSandboxChannel,
-  startSandboxChannel,
-  stopSandboxChannel,
-} from "./policy-channel-actions";
-
 type ChannelsRuntimeBridge = {
   sandboxChannelsAdd: (sandboxName: string, args?: string[]) => Promise<void>;
   sandboxChannelsRemove: (sandboxName: string, args?: string[]) => Promise<void>;
@@ -19,12 +12,20 @@ type ChannelsRuntimeBridge = {
   sandboxChannelsStop: (sandboxName: string, args?: string[]) => Promise<void>;
 };
 
-let runtimeBridgeFactory = (): ChannelsRuntimeBridge => ({
-  sandboxChannelsAdd: addSandboxChannel,
-  sandboxChannelsRemove: removeSandboxChannel,
-  sandboxChannelsStart: startSandboxChannel,
-  sandboxChannelsStop: stopSandboxChannel,
-});
+let runtimeBridgeFactory = (): ChannelsRuntimeBridge => {
+  const actions = require("./policy-channel-actions") as {
+    addSandboxChannel: ChannelsRuntimeBridge["sandboxChannelsAdd"];
+    removeSandboxChannel: ChannelsRuntimeBridge["sandboxChannelsRemove"];
+    startSandboxChannel: ChannelsRuntimeBridge["sandboxChannelsStart"];
+    stopSandboxChannel: ChannelsRuntimeBridge["sandboxChannelsStop"];
+  };
+  return {
+    sandboxChannelsAdd: actions.addSandboxChannel,
+    sandboxChannelsRemove: actions.removeSandboxChannel,
+    sandboxChannelsStart: actions.startSandboxChannel,
+    sandboxChannelsStop: actions.stopSandboxChannel,
+  };
+};
 
 export function setChannelsRuntimeBridgeFactoryForTest(
   factory: () => ChannelsRuntimeBridge,
