+++ src/lib/policy-mutate-cli-commands.ts
@@ -5,17 +5,21 @@
 
 import { Args, Command, Flags } from "@oclif/core";
 
-import { addSandboxPolicy, removeSandboxPolicy } from "./policy-channel-actions";
-
 type PolicyRuntimeBridge = {
   sandboxPolicyAdd: (sandboxName: string, args?: string[]) => Promise<void>;
   sandboxPolicyRemove: (sandboxName: string, args?: string[]) => Promise<void>;
 };
 
-let runtimeBridgeFactory = (): PolicyRuntimeBridge => ({
-  sandboxPolicyAdd: addSandboxPolicy,
-  sandboxPolicyRemove: removeSandboxPolicy,
-});
+let runtimeBridgeFactory = (): PolicyRuntimeBridge => {
+  const actions = require("./policy-channel-actions") as {
+    addSandboxPolicy: PolicyRuntimeBridge["sandboxPolicyAdd"];
+    removeSandboxPolicy: PolicyRuntimeBridge["sandboxPolicyRemove"];
+  };
+  return {
+    sandboxPolicyAdd: actions.addSandboxPolicy,
+    sandboxPolicyRemove: actions.removeSandboxPolicy,
+  };
+};
 
 export function setPolicyRuntimeBridgeFactoryForTest(factory: () => PolicyRuntimeBridge): void {
   runtimeBridgeFactory = factory;
